Major: correct logic when using middle_name:
If name strings composed of a first and a middle name sent to the API lead to contradicting gender depending on the connecting character (i.e. hyphen or empty space), then do not use the response; instead use the first name only to create a new query.

add placeholder to config